more changes for 64 bit computing warning errors.

+++ freex3d/src/lib/vrml_parser/CParseParser.c
@@ -2704,7 +2704,7 @@ void cParseErrorCurID(struct VRMLParser *me, char *str) {
 void cParseErrorFieldString(struct VRMLParser *me, char *str, const char *str2) {
 
 	char fw_outline[OUTLINELEN];
-	int str2len = strlen(str2);
+	size_t str2len = strlen(str2);
 
 	if (strlen(str) > FROMSRC) str[FROMSRC] = '\0';
 	strcpy(fw_outline,str);
